extensions/rcs_tacto/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "rcs_tacto"
version = "0.6.1"
description="RCS integration of tacto"
dependencies = [
    "rcs>=0.6.1",
    "omegaconf",
    # TODO: fork this repo into robot control stack
    "mujoco-tacto@git+https://github.com/utn-air/mujoco-tacto.git@main"
]
readme = "README.md"
maintainers = [
  { name = "Tobias Jülg", email = "tobias.juelg@utn.de" },
  { name = "Seongjin Bien", email = "seongjin.bien@utn.de" },
]
authors = [
  { name = "Tobias Jülg", email = "tobias.juelg@utn.de" },
  { name = "Seongjin Bien", email = "seongjin.bien@utn.de" },
]
requires-python = ">=3.10"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
